refactored app.js to rides.js refactored db callbacks to async/await updated schema to create table only if not exists added an external db file for rides
