+++ src/schemas.js
@@ -1,6 +1,6 @@
 module.exports = (db) => {
     const createRideTableSchema = `
-        CREATE TABLE Rides
+        CREATE TABLE IF NOT EXISTS Rides
         (
         rideID INTEGER PRIMARY KEY AUTOINCREMENT,
         startLat DECIMAL NOT NULL,
